remove-xmls.pl should remove a directory only when its not been modified for the amount of time given on the command line.

+++ dsc/trunk/presenter/cron/remove-xmls.pl
@@ -10,7 +10,8 @@ while (<>) {
 	chomp;
 	next if (/backlog/);
 	if (-d $_) {
-		rmdir $_;
+		my @sb = stat $_;
+		rmdir $_ if ($sb[9] < $cutoff);
 		next;
 	}
 	next unless (/\.xml/);
